requirements.txt
asgiref==3.8.1
Django==5.1.2
django-bootstrap5==24.3
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-precise-bbcode==1.2.16
django-ranged-response==0.2.0
django-redis==5.4.0
django-simple-captcha==0.6.0
djangorestframework==3.15.2
easy-thumbnails==2.10
pillow==11.1.0
psycopg2==2.9.10
redis==5.2.1
sqlparse==0.5.1
tzdata==2024.2
